24/07/26/ update receiptRawMaterial report

+++ com.ambassador.support.lib/ViewModel/ReceiptRawMaterialViewModel.cs
@@ -14,10 +14,10 @@ namespace com.ambassador.support.lib.ViewModel
         public string URNDate { get; set; }
         public string ProductCode { get; set; }
         public string ProductName { get; set; }
-        public decimal SmallQuantity { get; set; }
+        public double SmallQuantity { get; set; }
         public string SmallUomUnit { get; set; }
         public string DOCurrencyCode { get; set; }        
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string StorageName { get; set; }
         public string SupplierName { get; set; }
         public string Country { get; set; }
